backend/src/requirements.txt
# Django
Django==3.0.7

# PostgreSQL
psycopg2==2.8.5

# Fake data for testing
factory_boy==2.12.0

# Gunicorn
gunicorn==20.0.4

# Django jet
django-3-jet==1.0.8

# Utils
django-extensions==2.2.9
django-debug-toolbar==2.2
django-simple-history==2.10.0

# Make sure static file is well configured
whitenoise==5.1.0

# iPython
setuptools==47.1.1
